(#2) Subo endpoint y su respectiva logica para traer los establecimientos destacados

+++ AppNighPlanbackend/src/main/java/DAO/BaseDAO.java
@@ -13,7 +13,7 @@ public interface BaseDAO<T>{
     void actualizar(T t) throws DAOException;
     void eliminar(T t) throws DAOException;
     List<T>obtenerTodos() throws DAOException;
-    T obtener(Long id) throws DAOException;
+    T obtener(int id) throws DAOException;
     
     
 }
