+++ AppNighPlanbackend/src/main/java/DAO/GruposDAO.java
@@ -46,7 +46,7 @@ public class GruposDAO implements BaseDAO{
     }
 
     @Override
-    public Object obtener(Long id) {
+    public Object obtener(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
